feat: Shelly IO debug, event tracking, and TUI activity indicators
Shelly driver:
- Add DiscoverAll mode: auto-discover devices via MQTT announce messages
- Add thread-safe device registry (devicesMu RWMutex) and topicRoots map
- Add announce-based discovery (shellyDiscoveryHandler, onDeviceDiscovered)
- Implement IoDebugProvider (GetIoDebugSnapshot) and IoOutputToggler (ToggleOutput)
- Track output state changes via detectSwitchChanges (works for discovery-only
  mode, not just explicitly configured outputs)
- Track input event timestamps (inputLastEvent) updated on every NotifyEvent,
  even without a Button config subscription
- Add stateMu for outputLastChanged/inputLastEvent/prevSwitchStates maps
- Fix Subscribe to support bitmask of multiple event types

IO debug data layer:
- Add LastEvent time.Time to IoPointState and IoPointDebugState to distinguish
  explicit button-press events from plain state changes
- Wire Shelly into collectIoDebug in state_provider.go

TUI IO debug:
- Dynamic name column width: adapts to longest name in current list so Shelly
  device IDs (e.g. shellyplus1pm-ABCD1234:switch0) are not truncated
- Activity timer: counts up from 0s for 200s after any state change or event;
  sources are driver LastChanged, driver LastEvent, and TUI-observed state flip
- TUI-side change detection (detectIoStateChanges): starts timer when the
  rendered state dot changes colour, as a reliable fallback independent of
  driver timestamp population
- Event flash: state dot turns orange (ColorEvent=208) for 2s after an explicit
  input event, replacing the normal gray/yellow dot for that window

+++ drivers/shelly/get_events.go
@@ -9,14 +9,14 @@ import (
 
 type RawEvents struct {
 	Events []shellyEvent `json:"events"`
-	Ts     int           `json:"ts"`
+	Ts     float64       `json:"ts"`
 }
 
 type shellyEvent struct {
-	Ts        int    `json:"ts"`
-	Component string `json:"component"`
-	Id        uint   `json:"id"`
-	Event     string `json:"event"`
+	Ts        float64 `json:"ts"`
+	Component string  `json:"component"`
+	Id        uint    `json:"id"`
+	Event     string  `json:"event"`
 }
 
 type Event struct {
